+++ graal-js/src/com.oracle.truffle.js.test/commonjs/cycle_main.js
@@ -1,0 +1,13 @@
+/*
+ * Copyright (c) 2019, 2019, Oracle and/or its affiliates. All rights reserved.
+ * DO NOT ALTER OR REMOVE COPYRIGHT NOTICES OR THIS FILE HEADER.
+ *
+ * Licensed under the Universal Permissive License v 1.0 as shown at http://oss.oracle.com/licenses/upl.
+ */
+console.log('main starting at ' + __filename);
+exports.done = false;
+const other = require('./cycle_other.js');
+console.log('other.done = ' + other.done);
+exports.done = true;
+console.log('main done');
+42 + other.num;